chore: tidy up oauth exports
We should stick to one class per file ideally to make testing and
readability better.
